requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0
SpeechRecognition~=3.14.3
pydub~=0.25.1
prometheus_client~=0.22.1